refactor: standardize config model
