+++ src/pytest_drill_sergeant/plugin/base.py
@@ -15,7 +15,8 @@ from pydantic import BaseModel, Field
 if TYPE_CHECKING:
     from pathlib import Path
 
-    from pytest_drill_sergeant.core.models import Config, Finding
+    from pytest_drill_sergeant.core.config import DrillSergeantConfig as Config
+    from pytest_drill_sergeant.core.models import Finding
 
 logger = logging.getLogger(__name__)
