+++ tests/integration/test_config_integration.py
@@ -18,18 +18,13 @@ from pytest_drill_sergeant.core.config_manager import (
     initialize_config,
 )
 from pytest_drill_sergeant.core.config_validator import validate_configuration
-from pytest_drill_sergeant.core.models import RuleType
 from tests.constants import (
-    BIS_THRESHOLD_90,
     BIS_THRESHOLD_FAIL,
     BIS_THRESHOLD_WARN,
     BUDGET_ERROR,
     BUDGET_WARN,
     CUSTOM_THRESHOLD,
     CUSTOM_THRESHOLD_ALT,
-    LSP_PORT,
-    MAX_WORKERS,
-    SIMILARITY_THRESHOLD,
 )
 
 
@@ -365,34 +360,3 @@ budgets = { warn = 20, error = 5 }
         assert config.get_budget("warn") == BUDGET_WARN
         assert config.get_budget("error") == BUDGET_ERROR
         assert config.get_budget("nonexistent", 0) == 0
-
-    def test_configuration_to_base_config_conversion(self) -> None:
-        """Test conversion to base config."""
-        config = DrillSergeantConfig(
-            mode="strict",
-            persona="snoop_dogg",
-            sut_package="myapp",
-            budgets={"warn": 15, "error": 5},
-            enabled_rules={"aaa_comments"},
-            suppressed_rules=set(),
-            thresholds={"bis_threshold_warn": 90},
-            mock_allowlist={"custom.*"},
-        )
-
-        base_config = config.to_base_config()
-
-        assert base_config.mode == "strict"
-        assert base_config.persona == "snoop_dogg"
-        assert base_config.sut_package == "myapp"
-        assert base_config.fail_on_how is False
-        assert base_config.output_format == "terminal"
-        assert base_config.verbose is False
-        assert base_config.enabled_rules == [RuleType.AAA_COMMENT]
-        assert base_config.disabled_rules == []
-        assert base_config.similarity_threshold == SIMILARITY_THRESHOLD
-        assert base_config.bis_threshold == BIS_THRESHOLD_90
-        assert base_config.parallel_analysis is True
-        assert base_config.max_workers == MAX_WORKERS
-        assert base_config.cache_ast is True
-        assert base_config.lsp_enabled is False
-        assert base_config.lsp_port == LSP_PORT
